requirements.txt
flask==2.3.3
openpyxl==3.1.2
python-dotenv==1.0.0
Werkzeug==2.3.7
pywin32==306; platform_system == "Windows"
PyYAML==6.0.1
pandas==2.1.4
PyMySQL==1.1.0
requests==2.31.0
pycryptodome==3.17